requirements.txt
magic-pdf[full]>=0.8.0  # MinerU 的核心依赖
openai
python-docx>=1.1.2
pillow>=10.0.0
# 根据后续需求可以添加更多依赖
